prefs: enable "Send a Do-Not-Track header"
Set enable_do_not_track to true by default.

The web community has pointed out (in e.g.
https://news.ycombinator.com/item?id=9483958 ) that defaulting to
DNT=on is "a questionable plan". I agree with that feeling
that DNT becomes useless if everybody has it on.

However, Iridium is order of magnitudes away from "everybody"
currently. We shall reevaluate this in time again.

[The same ycombinator post also quotes section 6.3 of the
https://tools.ietf.org/html/draft-mayer-do-not-track-00#section-6
draft: "It MUST NOT transmit OPT-IN without explicit user consent."
This is ok, because the UA is not sending OPT-IN; it is sending
OPT-OUT.]